neo4j-mcp-server/pyproject.toml
[project]
name = "neo4j-mcp-server"
version = "0.1.0"
description = "Neo4j MCP Server on AWS Bedrock AgentCore with OAuth2 M2M authentication"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aws-cdk-lib>=2.170.0",
    "constructs>=10.0.0",
    "boto3>=1.42.0",
    "mcp>=1.10.0",
    "httpx>=0.28.0",
    "httpx-sse>=0.4.0",
    "requests>=2.31.0",
    "awscrt>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]
